Use page language rather than user language
We expect to use these messages in the context of a wiki page, where the
page language yields more consistent language use on wiki than the user
language. The full proposal for languages in Kartographer is available
here:
https://phabricator.wikimedia.org/T309166#8856291

We'll display page-internal messages in the page language rather than in
the user language (unless specified otherwise); this patch adjusts the
language fragment util for this, and creates a lang attribute method
that can be used to internationalize attributes (such as alt
attributes).

Bug: T263762
Bug: T309166
Change-Id: I50e54e83e5f032817ed510ebc2b86e5edd85f7a4