Mark more MapLinkAttributeGenerator properties as not nullable
These properties can only be null when the input is not valid, e.g.
when confronted with an invalid align="top". This is reported as an
error, and will continue to be reported as an error. But the
mechanism for this is the Status object. The fact these properties
can be null is never used for anything. I suggest to remove it to
make the interface of the generator class less complex.

This patch is an addition to I069f383, and hopefully easier to
review. It doesn't replace the other patch.

Bug: T263762
Change-Id: I087dab3b4f3854022df268cf7bad5c420e949bb0