Parsoid support for Kartographer version 1.0.0
This patch adds the support of Parsoid in Kartographer.

There is still a number of known failures that are related to the lack
of Parsoid l10n support for now. In practice, these will appear in
Parsoid renderings (including VE) when there is an error or when a
maplink does not have an explicit label (and its coordinates are
displayed instead). Additionally, until this is solved, we are avoiding
the question in the Parsoid path by setting the map display language to
English.

The Parsoid content is versioned to give the possibility of cache
rollbacks if necessary.

Bug: T263762
Change-Id: I9196d9dd931891f5f9570093235633d6ee5f0613