Process Kartographer content in attribute-embedded HTML
Selecting tags via CSS selector does not take into account the tags that
may be stashed in attributes. This transforms the existing
ParsoidDomProcessor to use a DOMTraverser internally so that relevant
attributes can also be processed.

The main use case of this patch is supporting Kartographer in indicators.
Temporary data is lost in this case (because it is not serialized in
indicators HTML), so we're moving the data transfer to an explicit
attribute as well.

Bug: T342887
Change-Id: I426b103a7b237dc23f64012b424818ecfa4e382d