Much faster tests by setting content language to qqx
I was quite surprised to see these popping up as some of the absolute
slowest tests on our CI, up to 2300ms for a single test case in one
case. It's not that bad locally but I can see where >90% of the time
is spend: Localization. That should be irrelevant for these tests.
They appear to work just fine with the raw, unlocalized qqx message
keys.

Locally this speeds up the entire Kartographer test suite from 2.3 to
1.9 seconds.

Bug: T225730
Change-Id: I51ee857c5d3e94ffb834a6bb388b671afb4691fc