Move "error" attribute into MapLinkAttributeGenerator
A few things need to change to make this possible:

1. The $counter is not stored as a property in two different
classes but in the MapTagArgumentValidator, along with the
user-provided text attribute.

2. Before it was helpful to know the difference between the
user-provided text and the automatically extracted counter. The
counter is guaranteed to be a trivial character or number and
doesn't need expensive wikitext parsing. This is now hidden by the
new method. Instead we use ctype_alnum() to decide if the text is
trivial or not. This covers all automatic counters (e.g. "A" or
"99").

Change-Id: I0b15ea151727343681aeae3538254536e61f0832