Remove most of the code for the obsolete WV Nearby
The feature is de-facto diabled for a few months now and the code
should be removed to free up ressources and improve performance.

There are a few endpoints still referenced by on wiki scripts that
will be kept as stub for now to avoid errors. But as mentioned
these endpoints also de-facto did nothing for some time.

Also makes the config setting obsolete.

For more information see the ticket attached.

The removal of the PruneCluster library code will follow as soon
as there are also no references left on wiki.

Bug: T332785
Change-Id: Ieaaa2bc91d96806e235b9bd55b17efaedf27fd67