Make ParsoidTagHandler stateless
\Wikimedia\Parsoid\Config\SiteConfig has a cache of ExtensionTagHandler
instances. This is problematic if said ExtensionTagHandlers have state
that does not get reset between multiple extension invocations - which
is the case for Kartographer ParsoidTagHandler (and consequently its
child classes ParsoidMapLink and ParsoidTagHandler).

This patch removes the state from ParsoidTagHandler by encapsulating the
fields in a data class that gets created on extension invocation and
passed as an argument where needed.

Bug: T347552
Change-Id: I0e106f67d309995dd2cc3094e29ecbbe82130ae8