MapTagArgumentValidator: Replace secondary property with getter
There is only one language that can be specified by the user via the
lang="…" attribute. That's why I always found it a bit confusing to
have two properties here.

"Resolved" is derived from "specified".

"Resolved" falls back to a default (one of two possible defaults,
depending on configuration). "Specified" does not.

Still "specified" is not the raw user input. It's validated and null
if the user input was invalid.

There is currently no code that needs the raw user input.

This patch introduces named methods for a few of the concepts listed
above.

A notable change in this patch is that the old code did the
validation after falling back to the page language. The new code
assumes that calling Language::getCode() must be valid.

Change-Id: If0d745ef3bb66f2d8f95275c52f7bef08fd615b7