Disallow whitespace-only group names in show="…" attribute
This was a flaw in the existing regular expression. There are two
ways to fix this: Allow it, but filter empty group names. However,
in all other cases empty group names are already disallowed. I feel
like it's better to stick to this behavior and not relax it.

I also normalized the regular expressions to make them easier to
read. In detail:
* (\w| ) is the same as [\w ]
* (|…) is the same as (…)?
* Instead of [a-zA-Z] we can make the pattern case-insensitive with
  the /i modifier.
* [_a-zA-Z0-9] is the same list of characters as in \w

The only relevant change is the additional plus at the end of one
of the \s*+. Such an extra plus stops backtracking.

Change-Id: Ia3aa0947bb97798fd806eadbe50d46a797858af5