Increase $wgKartographerStaticFullWidth to 1024
See the discussion in T354239. The value 1024 might be the maximum
we can use at the moment.

I believe it is easily possible to change this value however we like
without any preparation. The moment it is deployed pages with
<mapframe> tags will start using the new value, while old pages will
continue to use the old value for a little while before they get
refreshed. The Kartotherian server will happily deliver both
versions, I believe. Caching is done based on the URLs of the .png
images, and the width is part of these URLs.

You can manually test this: Go to any page with a width="full" map,
use the dev tools to edit the <img src="…"> and change the "800" in
the URL to "1024".

Server load shouldn't be a problem as such maps are rare.

Bug: T354239
Change-Id: I8e1775aa39caada6af8931a6167800c5cedf7e0f