Fix state bleeding from one <maplink> into the next
Caused by Iadc4697. Unfortunately we had no parser test that covered
this. The new test case covers this. With the bad code the color
"#abcdef" from the previous <maplink> bleeds into the next.

Note that this also disables a test in the nearby feature that
started failing randomly. See the other ticket attatched.

Bug: T355044
Bug: T355300
Change-Id: I12587a242f923ce279e4b97e6ab03f07366152d8