Replace deprecated PageConfig::getTitle()
I see core also uses the global Title::getPrefixedText() in the same
context, and not a service. Turns out it's not possible to use the
TitleFormatter service because that requires a core LinkTarget, but
PageConfig::getTitle returns a ParsoidLinkTarget. We would need to
change the TitleFormatter service first.

Should be good enough for now, as long as CI is happy.

Notice that PageConfig::getTitle returns the `getPrefixedText` string
including the namespace. Which means one of the places touched in
this patch added the namespace twice! Luckily it's not really used
for anything in TrackingCategories::resolveTrackingCategory.

Change-Id: If2a6e1434530518312bcb8323a92c7820aba93d1