Return separate Status from ParsoidTagHandler::parseTag
Main motivation here is to get rid of the hack where we manipulate
the status object inside the MapTagArgumentValidator. Instead return
the status as a separate thing.

While we could add a new $data->status property to the …Data
container I think it's equally valid to return a (shaped) array
instead. We can properly document this as an array shape now.

Change-Id: Ic3ed3a9e3eb9861424bf1659b000aff0f461e398