Add support for <mapframe alt="…"> alternative text
Critical for screenreader users. Same as in the [[File:…|alt=…]]
image syntax.

A user-provided alternative text is always best. Adding this as an
option doesn't block us from auto-generating alternative texts for
maps that don't have a user-provided one. Adding this later would
still be great, but needs more discussion.

Bug: T328137
Change-Id: I9aa55ebcf08f1ed78faf22775a0649de5d3270e1