Pass srcOffsets when parsing captions/texts with Parsoid
We were not passing srcOffsets to the ParsoidExtensionAPI when parsing
captions/texts in Parsoid tag handling. This could result in wrong
substring indices during template encapsulation, which could in turn
trigger a "bad UTF-8 at end of string" message, which allowed us to
detect this issue (fixed in this patch).

Bug: T362034
Change-Id: I8aafeb041d0d51b14e9445e5c7544f4fe92df132