Use ParserOutput::setNumericPageProperty()
This makes it clear that the page property is intended to be indexed for
a "top N" query, and avoids the (deprecated) use of setPageProperty()
for indexed page properties.

This does the same thing as 4f56f3eeb7e4d5186e52deac6a04b60754b05705
but in a hopefully more readable and less error-prone manner.

Depends-On: Ia94c192c429d0482c58467bed787fd2e0aca052f
Depends-On: I441699e7fe9827a5e06e4638ce88c685deb9b856
Followup-To: 4f56f3eeb7e4d5186e52deac6a04b60754b05705
Change-Id: I61f9f25c950015ee71ed204c71eb846b0e97efd8