Interpret tag arguments raw values
The attributes of extArgs passed to the tag API contain both a "vsrc"
value and a "v" value, which is a version of the value that has spaces
and newlines stripped.
For captions, we want to preserve these newlines, which may correspond
to wikitext containing lists; this is what this patch fixes.

Note the discrepancy on the legacy parser between behaviour through a
template or when calling the extension directly, as shown in tests. The
Parsoid behaviour is consistent between both calls (which seems like an
improvement). On production wikis, mapframe and maplink seem to be
called via templates more often than directly. Additionally, any user
who wished to display a list will have had to use the templated version
of the extension. Consequently, the discrepancy in the non-templated
version of this test is probably the right compromise (to be validated
in the next round of visual diff testing).

Bug: T359897
Change-Id: Ic8e26870dddd1b8d18d2a8bdb4559c9d7ce0ac23