Move TitleFormatter dependency to ParserContext
This is a direct follow-up for what I started in I77961dc. The
TitleFormatter dependency was really only used in a single, isolated
spot for a single purpose. It feels like it makes a lot of sense to
move it to the new ParserContext class. Now the two methods that
return the title and revision id both return very basic, scalar
values (string and int). Which is all we need.

Bug: T263762
Change-Id: I9ac2b8b47fe5dde2b6f826364d7077d9ac6e1f74