Switch to getPrefixedDBkey for <img> URLs
For some reason this code uses the human-readable text form since
2016, see Iba40a37. I believe this was a mistake. An older draft of
the same code was using getPrefixedDBkey just one year earlier, see
I965dc05.

This doesn't need to be human-readable but is part of the
<img src="…" srcset="…"> URLs for static map thumbnail images to be
generated by the Kartotherian service. Luckily both title forms work
just fine. Kartotherian doesn't really do anything with this title but
sends it back as part of a "mapdata" API request. This is a MediaWiki
query API which understands both title forms.

Changing this now can have an effect on caching. But that should be
fine, in my opinion:
* This is part of the <img …> HTML that's cached in the ParserCache.
  It will only be re-generated when there is a reason to re-parse a
  page.
* The Kartotherian server might indeed need to re-generate the image
  when the image URL changes.
* This won't affect pages where the two title forms are identical.
  Notably one-word page names without a space. This is approx. 1/3 of
  the high-traffic pages according to
  https://en.wikipedia.org/wiki/Wikipedia:Popular_pages
* This won't affect trivial maps without GeoJSON as these don't have
  the title in the image URL.

Change-Id: I7134e49e9a3ca67a6a019b693aa5153570e2b59b