Validate target language as well before using it
The target language is the content language of the page on most
normal wikis, but can be the user's interface language on
multi-lingual wikis like Wikimedia Commons. It appears like this
language code is not validated or normalized but provided as-is.
When the user uses …?uselang=ar--_or_ this will end in places like
the link to Special:Map.

Let's validate this as well.

This code is active for both parsers.

Bug: T288150
Change-Id: I275b3cd65dcff533292081131a7e765bce2ed204