Better placeholder color in the map dialog
This is the placeholder background color in the VisualEditor maps
dialog. It's only visible as long as the map tiles are being loaded.
While the currently used light gray is fine it's a color that
doesn't appear anywhere on the actual map tiles. This makes the
loading process more visible than it needs to be. When we instead
choose a color that's much closer to the incoming map tiles we can
make the loading process much less obvious.

When we look at https://maps.wikimedia.org there are 2 obvious
colors to choose from: the blue ocean or the bright yellow land.
Since almost all people live on land I figured it makes more sense
to choose that color.

Change-Id: I6a52a7ec30728744185e7304c1431fc134cec42b