modules: Combine overflow-x/overflow-y into overflow
modules/dialog/dialog.less
  28:2  ✖  Expected shorthand property "overflow"
declaration-block-no-redundant-longhand-properties

The overflow property is specified as one or two <overflow> keyword
values. If only one keyword is specified, both overflow-x and overflow-y
are set to the same value. If two keywords are specified, the first
value applies to overflow-x in the horizontal direction and the second
one applies to overflow-y in the vertical direction.

Change-Id: Ib95da8912d9be5587ca5e2fa55b8b9b0e0f645aa