Create new frame and clear DSR offset on extension tags
In T362034 we had introduced passing src offsets to the extension tags.
This breaks in multiple cases, including content that starts with a
space or content containing special characters. This patch instead
creates a new frame for the extension tag and clears the DSR, which
should avoid wrong information being passed around.

Bug: T373460
Change-Id: Iaa56d90b892996af7708ec3f7af7bd45d31176b4