Close dialog without confirmation if the user hasn't interacted
Recently we changed ve.ui.MWExtensionDialog to add `hasMeaningfulEdits`
which determines whether the user has interacted with a dialog and thus
allows the dialog to be closed without confirmation if they have made
no changes.

Kartographer keeps its state outside of the attributes and only copies
it over when asked to save. This change stores a copy of the initial
state in a hook that runs after the component has finished initializing.
The alternative would be to write a custom `hasMeaningfulEdits`, but
that would involve boilerplate bookkeeping any time a change is made to
add fields or change their defaults. This way is more robust.

This also disables the "save" action until the user has interacted when
inserting a new map. This is intended default behaviour of the parent
class if an inserted template has an initial state, and it seems like a
reasonable thing to inherit.

Bug: T334513
Change-Id: I5cb4b3d764d6064c1624e66ab0cc7557e2749fb4