Different file name for virtual vs. physical externalLinks.json
There is a physical file named externalLinks.json in the root
directory of this extension.

But the require( './externalLinks.json' ) call in the JavaScript
code doesn't refer to this file. It refers to a ResourceLoader
module that's declared via extension.json. The contents of the two
files are almost the same, except that the later is localized.

Also make the localization much more robust:
* Don't destroy existing "name" attributes. This makes it possible
  to hard-code names that don't need localization.
* Make it possible to disable messages and still show a non-empty
  fallback.

Bug: T374492
Change-Id: I27d2604e2c363a6beca6c4afa27426333bef43a8