parsoid <mapframe>: Fix handling of caption arguments
In addition, the ExtensionTagHandler says that "the extension tag
arguments should be treated as opaque objects and any necessary inspection
should be handled through the API."  To that end, use
ParsoidExtensionAPI::extArgsToArray() instead of the hacky converter
that was here before.

The test case 'List in mapframe caption, with mapframe defined
directly' documents that the 'text' attribute setting the mapframe
caption should be able to generate lists.  For legacy, this only works
when the content is protected by a transclusion, since attribute
whitespace is normalized.  For Parsoid, it was working previously
because we were processing the wikitext as tokens, not a string, and
so the list tokens were getting smuggled into the output even though
we were "supposed" to be normalizing attribute whitespace.

Make everything more sane by flagging the atttribute whitespace
handling for the `text` attribute explicitly as "keepspaces", which
lets us parse the wikitext caption as a string and get the desired
list output.  This is probably the best we can do with the current
Parsoid extension API which "eagerly evaluates" extension arguments.
The future PFragmentHandler API will let us better match legacy
behavior (ie allow list content only when using `{{#tag}}`), if we
decided to do so.

Depends-On: Idb1fe3a679e1b254f7da4ec6072685d58db3e1eb
Depends-On: I9d477a32e89f32187cc3505a730e4c701a8b6b73
Change-Id: I432288384340516be170db9a4741d56ed79854e1