Add a "parser" parameter to the ApiQueryMapData
Groups whose ID are generated by Kartographer rely on the hash sum of
the data, which can vary depending on whether the page is parsed via
Parsoid or the legacy parser. In this patch, we add a "parser"
parameter to the API to be able to get the groups from a Parsoid-parsed
page, and hence be able to render them on static maps as the API point
gets called.
The parameter is not used yet by the calling API so that its activation
can be handled separately.

Bug: T383328
Change-Id: If5537453a929033e4d4086c06b79d2acde14fa96