Avoid appending an empty Document Fragment
The example case from the log in the task is a caption consisting
entirely of a template, so that the caption string is not empty.
However, the template expansion returns nothing.  Normally, an empty
template response would still be wrapped but the mapframe itself is
nested in a template, so template wrapping is disabled.

The change to maplink is just by inspection.

Bug: T402548
Change-Id: Ia7bf2e148e854f411e8159a852079d51bfabb84f